Move services' classes to separated files
