+++ microchat/services/general_exceptions.py
@@ -1,0 +1,13 @@
+from abc import ABC
+
+
+class ServiceError(ABC, Exception):
+    pass
+
+
+class AccessDenied(ServiceError):
+    pass
+
+
+class DoesNotExists(ServiceError):
+    pass
